Move chat completion service to a common location
