+++ python/samples/concepts/chat_completion/simple_chatbot_logit_bias.py
@@ -2,7 +2,7 @@
 
 import asyncio
 
-from samples.concepts.chat_completion.chat_completion_services import (
+from samples.concepts.setup.chat_completion_services import (
     Services,
     get_chat_completion_service_and_request_settings,
 )
