+++ python/samples/concepts/setup/chat_completion_services.py
